+++ plugin_ide.ui/src/com/github/rustdt/ide/ui/navigator/RustNavigatorContentProvider.java
@@ -10,7 +10,14 @@
  *******************************************************************************/
 package com.github.rustdt.ide.ui.navigator;
 
+import org.eclipse.core.resources.IProject;
+
+import com.github.rustdt.ide.core.cargomodel.RustBundleInfo;
+
+import melnorme.lang.ide.core.LangCore_Actual;
+import melnorme.lang.ide.core.project_model.view.DependenciesContainer;
 import melnorme.lang.ide.ui.views.AbstractNavigatorContentProvider;
+import melnorme.utilbox.collections.ArrayList2;
 
 public class RustNavigatorContentProvider extends AbstractNavigatorContentProvider {
 	
@@ -23,6 +30,13 @@ public class RustNavigatorContentProvider extends AbstractNavigatorContentProvid
 	@Override
 	protected LangNavigatorSwitcher_GetChildren getChildren_switcher() {
 		return new LangNavigatorSwitcher_GetChildren() {
+			@Override
+			public void addFirstProjectChildren(IProject project, ArrayList2<Object> projectChildren) {
+				RustBundleInfo projectInfo = LangCore_Actual.getBundleModel().getProjectInfo(project);
+				if(projectInfo != null) {
+					projectChildren.add(new DependenciesContainer(projectInfo, project));
+				}
+			}
 		};
 	}