+++ documentation/ChangeLog.md
@@ -1,7 +1,8 @@
 ## RustDT release ChangeLog
 
 ### (NextVersion)
- * Added custom Project Explorer icons for Rust crate elements such as `src`, `test`, `target` folders, and the `Cargo.toml` file.
+ * Added specific Project Explorer icons for Rust crate elements such as `src`, `test`, `target` folders, and the `Cargo.toml` file.
+ * Added Project Explorer element with Cargo unresolved dependencies.
  * Fixed: Properly parse Racer "END" line syntax, don't show error when this occurs.
  * Fixed #41: Don't show error dialog when Racer fails. Only the editor status line is updated now.
  * Fixed: Occasional AssertionFailure when creating new projects in nested locations (project would not show up in Explorer).